+++ source/Examples/Resources/DnsServerEDns/1-SetCacheTimeout_Config.ps1
@@ -2,7 +2,7 @@
 
 .VERSION 1.0.0
 
-.GUID e3aeafd4-b41a-48e0-b9be-9b5c01f904d3
+.GUID 95d58494-7f07-4893-9afd-df5afc2a0509
 
 .AUTHOR DSC Community
 
@@ -25,7 +25,7 @@
 .EXTERNALSCRIPTDEPENDENCIES
 
 .RELEASENOTES
-Updated author, copyright notice, and URLs.
+First version.
 
 .PRIVATEDATA 2016-Datacenter,2016-Datacenter-Server-Core
 
@@ -35,20 +35,20 @@ Updated author, copyright notice, and URLs.
 
 <#
     .DESCRIPTION
-        This configuration will enable scavenging on the DNS server, using
-        the default interval values.
+        This configuration will change the cache timeout for
+        extension mechanisms for DNS (EDNS) on the DNS server.
 #>
 
-Configuration EnableScavenging_Config
+Configuration SetCacheTimeout_Config
 {
     Import-DscResource -ModuleName 'xDnsServer'
 
     Node localhost
     {
-        DnsServerScavenging 'EnableScavenging'
+        DnsServerEDns 'SetCacheTimeout'
         {
-            DnsServer       = 'localhost'
-            ScavengingState = $true
+            DnsServer    = 'localhost'
+            CacheTimeout = '0.00:15:00'
         }
     }
 }